requirements.txt
# ============================================
# Requirements FINAL para Python 3.10.11
# Windows & Linux - FUNCIONAL ✅
# ============================================

# Core AI/ML - Transcripción y IA
openai==1.109.1
torch==2.5.1
ctranslate2==4.5.0
av==12.3.0  # PyAV - Versión pre-compilada que funciona en Windows
faster-whisper==1.1.0  # Versión compatible con av 12.3.0

# Langchain - Para integración con GPT-4
langchain==0.3.27
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langsmith==0.4.31

# Video Processing - Procesamiento de video
moviepy==1.0.3
opencv-python==4.7.0.72
imageio==2.37.0
imageio-ffmpeg==0.6.0
pillow==11.2.1
proglog==0.1.12
decorator==4.4.2

# Audio Processing - Procesamiento de audio
pydub==0.25.1
webrtcvad-wheels==2.0.14

# YouTube Download - Descarga de videos
pytubefix==9.1.1

# FFmpeg Python bindings
ffmpeg-python==0.2.0

# HTTP & Networking
aiohttp==3.12.11
aiohappyeyeballs==2.6.1
aiosignal==1.3.2
async-timeout==4.0.3
httpx==0.28.1
httpcore==1.0.9
requests==2.32.3
urllib3==2.4.0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10

# Data Processing & Validation
pydantic==2.11.5
pydantic-core==2.33.2
numpy==1.26.0
networkx==3.4.2

# Utilities
python-dotenv==1.0.1
tqdm==4.67.1
coloredlogs==15.0.1
humanfriendly==10.0
packaging==25.0
filelock==3.18.0
fsspec==2025.5.1
pyyaml==6.0.2
regex==2025.9.18

# Hugging Face & Model Management
huggingface-hub==0.32.4
tokenizers==0.15.2
tiktoken==0.11.0

# Additional Dependencies
attrs==25.3.0
anyio==4.9.0
sniffio==1.3.1
h11==0.16.0
frozenlist==1.6.2
multidict==6.4.4
yarl==1.20.0
propcache==0.3.1
greenlet==3.2.4
sqlalchemy==2.0.43
jinja2==3.1.6
markupsafe==3.0.2
orjson==3.11.3
protobuf==6.31.1
flatbuffers==25.2.10
annotated-types==0.7.0
typing-extensions==4.14.0
typing-inspection==0.4.1
exceptiongroup==1.3.0
distro==1.9.0
tenacity==9.1.2
jsonpatch==1.33
jsonpointer==3.0.0
jiter==0.10.0
mpmath==1.3.0
sympy==1.13.1
future==1.0.0
zstandard==0.25.0
requests-toolbelt==1.0.0

# ONNX Runtime (CPU only)
onnxruntime==1.20.1
